+++ buildSrc/build.gradle
@@ -1,0 +1,29 @@
+buildscript {
+  // The "buildSrc" lies in its own project,
+  // independent of the remaining modules.
+  // Therefore, manually import the contents of our gradle.properties
+  // to be able to reference version constants.
+  def properties = new Properties()
+  file("../gradle.properties").withReader { properties.load(it) }
+  properties.each { key, value ->
+    ext[key] = value
+  }
+
+  repositories {
+    jcenter()
+  }
+  dependencies {
+    classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$KOTLIN_VERSION"
+  }
+}
+
+apply plugin: "kotlin"
+
+repositories {
+  mavenCentral()
+}
+
+dependencies {
+  compile gradleApi()
+  compile "org.jetbrains.kotlin:kotlin-stdlib:$KOTLIN_VERSION"
+}
