+++ gradle.properties
@@ -40,5 +40,5 @@ JUNIT_PLATFORM_VERSION                      = 1.0.2
 JUNIT_JUPITER_VERSION                       = 5.0.2
 JUNIT_VINTAGE_VERSION                       = 4.12.2
 JUNIT4_VERSION                              = 4.12
-SPOCK_VERSION                               = 1.2-groovy-2.4-SNAPSHOT
+SPOCK_VERSION                               = 1.1-groovy-2.4
 APACHE_COMMONS_VERSION                      = 2.6
