+++ tests/common.gradle
@@ -1,0 +1,59 @@
+import de.mannodermaus.gradle.plugins.junit5.WriteClasspathResource
+import org.apache.tools.ant.filters.ReplaceTokens
+
+apply plugin: "groovy"
+apply plugin: "java-gradle-plugin"
+apply plugin: "java-library"
+apply plugin: "idea"
+apply plugin: "jacoco"
+
+// Add custom dependency configurations
+configurations {
+  functionalTest {
+    description = "Local dependencies used for compiling & running tests source code " +
+        "inside of Gradle functional tests"
+  }
+}
+
+processTestResources {
+  def tokens = [COMPILE_SDK_VERSION: COMPILE_SDK_VERSION,
+                BUILD_TOOLS_VERSION: BUILD_TOOLS_VERSION,
+                MIN_SDK_VERSION    : MIN_SDK_VERSION,
+                TARGET_SDK_VERSION : TARGET_SDK_VERSION]
+
+  inputs.properties(tokens)
+
+  from(sourceSets.test.resources.srcDirs) {
+    include '**/testenv.properties'
+    filter(ReplaceTokens, tokens: tokens)
+  }
+}
+
+tasks.withType(WriteClasspathResource).all {
+  processTestResources.finalizedBy it
+}
+
+dependencies {
+  // Utilities for assertions & tests in general
+  testCompile "commons-io:commons-io:$APACHE_COMMONS_VERSION"
+  testCompile "commons-lang:commons-lang:$APACHE_COMMONS_VERSION"
+
+  // Dependent Plugins
+  // (Android Plugin is used only for compilation,
+  // the actual dependency is provided by the
+  // testAgp2x/testAgp3x modules respectively)
+  testCompileOnly "com.android.tools.build:gradle:$ANDROID_PLUGIN_3X_VERSION"
+  testCompile "org.jetbrains.kotlin:kotlin-gradle-plugin:$KOTLIN_VERSION"
+
+  // Test Execution
+  testCompile "junit:junit:$JUNIT4_VERSION"
+  testCompile("org.spockframework:spock-core:$SPOCK_VERSION") {
+    exclude group: "org.codehaus.groovy"
+  }
+
+  // Compilation of local classpath for functional tests
+  functionalTest "junit:junit:$JUNIT4_VERSION"
+  functionalTest "org.junit.jupiter:junit-jupiter-api:$JUNIT_JUPITER_VERSION"
+  functionalTest "org.junit.jupiter:junit-jupiter-engine:$JUNIT_JUPITER_VERSION"
+  functionalTest "org.junit.platform:junit-platform-console:$JUNIT_PLATFORM_VERSION"
+}
