+++ tests/testAgp2x/src/test/groovy/de/mannodermaus/gradle/plugins/junit5/agp2/AGP2Something.groovy
@@ -1,0 +1,7 @@
+package de.mannodermaus.gradle.plugins.junit5.agp2
+
+import de.mannodermaus.gradle.plugins.junit5.BaseSomething
+
+class AGP2Something extends BaseSomething {
+
+}
