+++ tests/testCommon/src/test/groovy/de/mannodermaus/gradle/plugins/junit5/BaseSomething.groovy
@@ -1,0 +1,8 @@
+package de.mannodermaus.gradle.plugins.junit5
+
+import spock.lang.Specification
+
+abstract class BaseSomething extends Specification {
+
+  def "Something should work"() {}
+}
