+++ tests/testCommon/src/test/groovy/de/mannodermaus/gradle/plugins/junit5/util/StringUtils.groovy
