+++ tests/testCommon/src/test/resources/de/mannodermaus/gradle/plugins/junit5/testenv.properties
